Update project settings for Xcode 8.2.1 (from 5.1...) Update to distrib for iOS 9.x and later Update to fix BTLE manager support Fix compiler warnings

+++ LowEnergyDemo/LowEnergyDemo/DVC-TISensorTag/LEDTISensorTag.m
@@ -438,7 +438,7 @@ NSString *kPERIPHERAL_SCAN_ENDED_NOTIFICATION = @"PANEL_NOTIFICATION_PERIPHERAL_
     ISPPeripheralTriadParameter *infoObject = notification.object;
     if(infoObject.error != nil)
     {
-        DLog(@"-- !!! ERROR(%d) %@", infoObject.error.code, infoObject.error.localizedDescription);
+        DLog(@"-- !!! ERROR(%ld) %@", (long)infoObject.error.code, infoObject.error.localizedDescription);
     }
 
     NSString *strArrivingChrstcUUID = infoObject.characteristic.UUID.UUIDString;
